requirements.txt
torch
torchvision
pillow
matplotlib
datasets
numpy
